services/violations/requirements.txt
fastapi
gunicorn
uvicorn[standard]
google-cloud-bigquery
google-cloud-pubsub
google-cloud-storage

